+++ src/pydbsec/__init__.py
@@ -25,4 +25,4 @@ __all__ = [
     "APIError",
 ]
 
-__version__ = "0.5.0"
+__version__ = "0.5.2"